implement user balance retrieval in PaymentService
