+++ packages/apps/job-launcher/server/src/modules/user/user.controller.ts
@@ -18,6 +18,7 @@ import { RequestWithUser } from '../../common/types';
 import { ErrorUser } from '../../common/constants/errors';
 import { UserBalanceDto } from './user.dto';
 import { ApiKey } from '../../common/decorators';
+import { PaymentService } from '../payment/payment.service';
 
 @ApiBearerAuth()
 @UseGuards(JwtAuthGuard)
@@ -27,7 +28,10 @@ import { ApiKey } from '../../common/decorators';
 export class UserController {
   private readonly logger = new Logger(UserController.name);
 
-  constructor(private readonly userService: UserService) {}
+  constructor(
+    private readonly userService: UserService,
+    private readonly paymentService: PaymentService,
+  ) {}
 
   @ApiOperation({
     summary: 'Get user balance',
@@ -47,7 +51,7 @@ export class UserController {
     @Request() req: RequestWithUser,
   ): Promise<UserBalanceDto> {
     try {
-      return this.userService.getBalance(req.user.id);
+      return this.paymentService.getUserBalance(req.user.id);
     } catch (e) {
       this.logger.log(
         e.message,
