+++ packages/apps/job-launcher/server/src/modules/user/user.service.spec.ts
@@ -2,10 +2,8 @@ import { createMock } from '@golevelup/ts-jest';
 import { HttpService } from '@nestjs/axios';
 import { ConfigService } from '@nestjs/config';
 import { Test } from '@nestjs/testing';
-import { PaymentCurrency } from '../../common/enums/payment';
 import { UserStatus, UserType } from '../../common/enums/user';
-import { PaymentService } from '../payment/payment.service';
-import { UserBalanceDto, UserCreateDto } from './user.dto';
+import { UserCreateDto } from './user.dto';
 import { UserEntity } from './user.entity';
 import { UserRepository } from './user.repository';
 import { UserService } from './user.service';
@@ -14,7 +12,6 @@ jest.mock('@human-protocol/sdk');
 
 describe('UserService', () => {
   let userService: UserService;
-  let paymentService: PaymentService;
   let userRepository: UserRepository;
 
   beforeAll(async () => {
@@ -24,7 +21,6 @@ describe('UserService', () => {
       providers: [
         UserService,
         { provide: UserRepository, useValue: createMock<UserRepository>() },
-        { provide: PaymentService, useValue: createMock<PaymentService>() },
         { provide: ConfigService, useValue: mockConfigService },
         { provide: HttpService, useValue: createMock<HttpService>() },
       ],
@@ -32,11 +28,6 @@ describe('UserService', () => {
 
     userService = moduleRef.get<UserService>(UserService);
     userRepository = moduleRef.get(UserRepository);
-    paymentService = moduleRef.get(PaymentService);
-  });
-
-  afterEach(() => {
-    jest.resetAllMocks();
   });
 
   describe('create', () => {
@@ -93,65 +84,5 @@ describe('UserService', () => {
       expect(result).toBe(null);
       expect(userRepository.findByEmail).toHaveBeenCalledWith(email);
     });
-  });
-
-  describe('getBalance', () => {
-    it('should return the correct balance with currency for a user', async () => {
-      const userId = 1;
-      const expectedBalance: UserBalanceDto = {
-        balances: [
-          { currency: PaymentCurrency.USD, amount: 100 },
-          { currency: PaymentCurrency.HMT, amount: 50 },
-          { currency: PaymentCurrency.USDT, amount: 0 },
-        ],
-        totalUsdAmount: 150,
-      };
-
-      jest
-        .spyOn(paymentService, 'getUserBalanceByCurrency')
-        .mockResolvedValueOnce(100)
-        .mockResolvedValueOnce(50)
-        .mockResolvedValue(0);
-      jest
-        .spyOn(paymentService, 'convertToUSD')
-        .mockResolvedValueOnce(100)
-        .mockResolvedValueOnce(50)
-        .mockResolvedValue(0);
-
-      const balance = await userService.getBalance(userId);
-
-      expect(balance).toEqual(expectedBalance);
-      expect(paymentService.getUserBalanceByCurrency).toHaveBeenCalledTimes(3);
-      expect(paymentService.convertToUSD).toHaveBeenCalledTimes(3);
-    });
-  });
-
-  it('should return zero balance for new users', async () => {
-    const userId = 1;
-    const expectedBalance: UserBalanceDto = {
-      balances: [
-        { currency: PaymentCurrency.USD, amount: 0 },
-        { currency: PaymentCurrency.HMT, amount: 0 },
-        { currency: PaymentCurrency.USDT, amount: 0 },
-      ],
-      totalUsdAmount: 0,
-    };
-
-    jest
-      .spyOn(paymentService, 'getUserBalanceByCurrency')
-      .mockResolvedValueOnce(0)
-      .mockResolvedValueOnce(0)
-      .mockResolvedValue(0);
-    jest
-      .spyOn(paymentService, 'convertToUSD')
-      .mockResolvedValueOnce(0)
-      .mockResolvedValueOnce(0)
-      .mockResolvedValue(0);
-
-    const balance = await userService.getBalance(userId);
-
-    expect(balance).toEqual(expectedBalance);
-    expect(paymentService.getUserBalanceByCurrency).toHaveBeenCalledTimes(3);
-    expect(paymentService.convertToUSD).toHaveBeenCalledTimes(3);
   });
 });
