+++ packages/apps/job-launcher/server/src/modules/user/user.service.ts
@@ -3,20 +3,14 @@ import * as bcrypt from 'bcrypt';
 
 import { UserEntity } from './user.entity';
 import { UserStatus, UserType } from '../../common/enums/user';
-import { UserBalanceDto, UserCreateDto, CurrencyBalanceDto } from './user.dto';
+import { UserCreateDto } from './user.dto';
 import { UserRepository } from './user.repository';
 import { ValidatePasswordDto } from '../auth/auth.dto';
-import { PaymentService } from '../payment/payment.service';
-import { PaymentCurrency } from '../../common/enums/payment';
-import { add } from '../../common/utils/decimal';
 
 @Injectable()
 export class UserService {
   private HASH_ROUNDS = 12;
-  constructor(
-    private userRepository: UserRepository,
-    private readonly paymentService: PaymentService,
-  ) {}
+  constructor(private userRepository: UserRepository) {}
 
   public async create(dto: UserCreateDto): Promise<UserEntity> {
     const newUser = new UserEntity();
@@ -47,28 +41,5 @@ export class UserService {
   ): Promise<UserEntity> {
     userEntity.password = bcrypt.hashSync(data.password, this.HASH_ROUNDS);
     return this.userRepository.updateOne(userEntity);
-  }
-
-  public async getBalance(userId: number): Promise<UserBalanceDto> {
-    const balances: CurrencyBalanceDto[] = [];
-    let totalUSDAmount = 0;
-
-    for (const currency of Object.values(PaymentCurrency)) {
-      const amount = await this.paymentService.getUserBalanceByCurrency(
-        userId,
-        currency,
-      );
-      const amountInUSD = await this.paymentService.convertToUSD(
-        amount,
-        currency,
-      );
-      totalUSDAmount = add(totalUSDAmount, amountInUSD);
-      balances.push({ currency, amount });
-    }
-
-    return {
-      balances,
-      totalUsdAmount: totalUSDAmount,
-    };
   }
 }
